fix circular dependency in text-input
